refactor: Make changes in folder structure
